% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myGWAS_fit}
\alias{myGWAS_fit}
\title{Fit a stan model.}
\usage{
myGWAS_fit(model = "dummy", method = "dummy", type = "dummy",
  cov = NULL, geno = NULL, pheno = NULL, prior_non = 1, prior_inf = 0,
  iter = 1000, warmup = 200, chains = 3, cores = 3)
}
\arguments{
\item{model}{A stan model of rstan:stanmodel class.}

\item{method}{Method to estimate model parameters ("optimizing", or "sampling").}

\item{type}{Variable type of the response variable ("linear", "binary", "categorical").}

\item{cov}{A matrix of model covariates.}

\item{geno}{A vector of variant genotype data (0, 1, 2).}

\item{pheno}{A vector of response variable data.}
}
\value{
A list of model fit.
}
\description{
Fit a stan model.
}
\examples{
myGWAS_fit(mymodel, rebuild = F)
data("sample", package = "bayes.glmm")
pheno = data$pheno
geno = data$geno
cov = data$cov
L <- t(chol(data$K))
mymodel = stan_models(model = "glm")
y1 = myGWAS_fit(model = mymodel, method = "optimizing", type = "categorical", cov = cov, geno = geno[c(1, 3), ], pheno = pheno)
y2 = myGWAS_fit(model = mymodel, method = "sampling", type = "categorical", cov = cov, geno = geno[c(1, 3), ], pheno = pheno)
}
\seealso{
\code{\link{rstan package}} which this function wraps
}
\author{
Xulong Wang, \email{xwang@jax.org}
}
