% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{stan_models}
\alias{stan_models}
\title{Choose a stan model ("glm" or "glmm").}
\usage{
stan_models(model = "dummy", rebuild = F)
}
\arguments{
\item{model}{model type ("glm" or "glmm)}

\item{rebuild}{whether to rebuild the stan model from source (TRUE or FALSE)}
}
\value{
A stan model in rstan::stanmodel class
}
\description{
Choose a stan model ("glm" or "glmm").
}
\examples{
stan_models("glm", rebuild = F)
stan_models("glmm", rebuild = F)
}
\seealso{
\code{\link{extdata/*.stan}} for source code.
}
\author{
Xulong Wang, \email{xwang@jax.org}
}
